+++ catches/views.py
@@ -8,16 +8,23 @@ import csv
 import io
 import json
 import os
+import numpy as np
+import pandas as pd
+from sklearn.preprocessing import MinMaxScaler
+from sklearn.linear_model import LinearRegression
+from sklearn.neural_network import MLPRegressor
+from sklearn.metrics import mean_squared_error
+from statsmodels.tsa.arima.model import ARIMA
+import matplotlib.pyplot as plt
+from platypus import NSGAIII, Problem, Real
+import warnings
+from datetime import datetime
+
+warnings.filterwarnings('ignore')
+
 from .models import FishCatch, CatchDetail
 from .serializers import FishCatchSerializer, CatchDetailSerializer, FishCatchWithDetailsSerializer
 from drf_spectacular.utils import extend_schema, extend_schema_view
-
-# Import PNBP prediction module
-try:
-    from pnbp_predictions import PNBPredictor, save_pnbp_predictions
-except ImportError:
-    PNBPredictor = None
-    save_pnbp_predictions = None
 
 @extend_schema_view(
     list=extend_schema(
@@ -547,6 +554,351 @@ class CatchDetailViewSet(viewsets.ModelViewSet):
 
 
 # -----------------------------------------------------------------------------------
+# PNBP Prediction Classes
+# -----------------------------------------------------------------------------------
+
+class PNBPRegressionPredictor:
+    """Prediktor PNBP menggunakan Analisis Regresi"""
+
+    @staticmethod
+    def predict(historical_data):
+        """
+        Analisis Regresi untuk memprediksi PNBP berdasarkan hubungan antara
+        variabel produksi, harga ikan, biaya operasional, dan kapasitas kapal.
+        """
+        if len(historical_data) < 3:
+            return PNBPRegressionPredictor._fallback_prediction(historical_data)
+
+        # Prepare features: volume, GT (simulated), operational costs (simulated)
+        features = []
+        targets = []
+
+        for item in historical_data:
+            volume = item.get('total_volume', 100)
+            gt = item.get('gt_kapal', 60)  # Default GT
+            operational_cost = volume * 0.1  # Simulated operational cost
+            features.append([volume, gt, operational_cost])
+            targets.append(item.get('total_pnbp', 0))
+
+        try:
+            # Train regression model
+            model = LinearRegression()
+            model.fit(features, targets)
+
+            # Predict for next period (assume average values)
+            avg_volume = np.mean([f[0] for f in features])
+            avg_gt = np.mean([f[1] for f in features])
+            avg_cost = np.mean([f[2] for f in features])
+
+            prediction = model.predict([[avg_volume, avg_gt, avg_cost]])[0]
+            return max(0, prediction)
+        except Exception as e:
+            print(f"Regression prediction error: {e}")
+            return PNBPRegressionPredictor._fallback_prediction(historical_data)
+
+    @staticmethod
+    def _fallback_prediction(historical_data):
+        """Fallback prediction using simple average"""
+        if not historical_data:
+            return 0
+        return np.mean([item.get('total_pnbp', 0) for item in historical_data])
+
+
+class PNBPNeuralNetworkPredictor:
+    """Prediktor PNBP menggunakan Jaringan Saraf"""
+
+    @staticmethod
+    def predict(historical_data):
+        """
+        Jaringan Saraf untuk mendeteksi pola non-linear dan tren jangka panjang pada data historis PNBP.
+        """
+        if len(historical_data) < 5:
+            return PNBPNeuralNetworkPredictor._fallback_prediction(historical_data)
+
+        # Prepare time series data
+        pnbp_values = [item.get('total_pnbp', 0) for item in historical_data]
+
+        # Create features: lagged values
+        X = []
+        y = []
+        for i in range(3, len(pnbp_values)):
+            X.append(pnbp_values[i-3:i])
+            y.append(pnbp_values[i])
+
+        if len(X) < 2:
+            return PNBPNeuralNetworkPredictor._fallback_prediction(historical_data)
+
+        try:
+            # Train neural network
+            model = MLPRegressor(hidden_layer_sizes=(50, 25), max_iter=1000, random_state=42)
+            model.fit(X, y)
+
+            # Predict next value
+            last_sequence = pnbp_values[-3:]
+            prediction = model.predict([last_sequence])[0]
+            return max(0, prediction)
+        except Exception as e:
+            print(f"Neural network prediction error: {e}")
+            return PNBPNeuralNetworkPredictor._fallback_prediction(historical_data)
+
+    @staticmethod
+    def _fallback_prediction(historical_data):
+        """Fallback prediction using simple average"""
+        if not historical_data:
+            return 0
+        return np.mean([item.get('total_pnbp', 0) for item in historical_data])
+
+
+class PNBPOptimizationPredictor:
+    """Prediktor PNBP menggunakan Optimasi Multi-Objektif NSGA-III"""
+
+    @staticmethod
+    def predict(historical_data, ships_data=None):
+        """
+        Optimasi Multi-Objektif untuk mencari kombinasi variabel optimal
+        yang memaksimalkan potensi pendapatan dengan tetap menjaga keberlanjutan.
+        """
+        if not ships_data or len(ships_data) == 0:
+            return PNBPOptimizationPredictor._fallback_prediction(historical_data)
+
+        num_ships = len(ships_data)
+
+        def pnbp_objective_function(vars):
+            # vars represent efficiency factors for each ship
+            efficiencies = np.array(vars)
+
+            # Calculate potential PNBP based on efficiency and historical data
+            avg_historical_pnbp = np.mean([item.get('total_pnbp', 0) for item in historical_data])
+            base_pnbp = avg_historical_pnbp * 0.8  # Conservative base
+
+            # Efficiency-weighted PNBP
+            efficiency_bonus = np.sum(efficiencies) * avg_historical_pnbp * 0.2
+
+            total_pnbp = base_pnbp + efficiency_bonus
+
+            # Objectives: maximize PNBP, minimize resource usage variance, minimize environmental impact
+            resource_variance = np.var(efficiencies)
+            environmental_impact = np.sum(efficiencies > 1.2) * 100  # Penalty for over-efficiency
+
+            return [-total_pnbp, resource_variance, environmental_impact]
+
+        try:
+            problem = Problem(num_ships, 3)
+            problem.types[:] = Real(0.5, 1.5)
+            problem.function = pnbp_objective_function
+
+            algorithm = NSGAIII(problem, divisions_outer=12, divisions_inner=2)
+            algorithm.run(100)
+
+            if not algorithm.result:
+                return PNBPOptimizationPredictor._fallback_prediction(historical_data)
+
+            # Use the best solution
+            best_solution = algorithm.result[0]
+            predicted_pnbp = -best_solution.objectives[0]  # Negate because we maximized negative
+
+            return max(0, predicted_pnbp)
+        except Exception as e:
+            print(f"Optimization prediction error: {e}")
+            return PNBPOptimizationPredictor._fallback_prediction(historical_data)
+
+    @staticmethod
+    def _fallback_prediction(historical_data):
+        """Fallback prediction using simple average"""
+        if not historical_data:
+            return 0
+        return np.mean([item.get('total_pnbp', 0) for item in historical_data])
+
+
+class PNBPTimeSeriesPredictor:
+    """Prediktor PNBP menggunakan Peramalan Deret Waktu"""
+
+    @staticmethod
+    def predict(historical_data):
+        """
+        Peramalan Deret Waktu untuk memprediksi nilai PNBP di periode mendatang
+        berdasarkan pola musiman dan fluktuasi tahunan.
+        """
+        if len(historical_data) < 5:
+            return PNBPTimeSeriesPredictor._fallback_prediction(historical_data)
+
+        # Extract PNBP time series
+        pnbp_values = [item.get('total_pnbp', 0) for item in historical_data]
+
+        try:
+            # Fit ARIMA model (p=1, d=1, q=1) - common starting point
+            model = ARIMA(pnbp_values, order=(1, 1, 1))
+            model_fit = model.fit()
+
+            # Forecast next value
+            forecast = model_fit.forecast(steps=1)[0]
+            return max(0, forecast)
+        except Exception as e:
+            print(f"Time series prediction error: {e}")
+            return PNBPTimeSeriesPredictor._fallback_prediction(historical_data)
+
+    @staticmethod
+    def _fallback_prediction(historical_data):
+        """Fallback prediction using exponential smoothing"""
+        if not historical_data:
+            return 0
+
+        pnbp_values = [item.get('total_pnbp', 0) for item in historical_data]
+        if len(pnbp_values) >= 2:
+            alpha = 0.3
+            smoothed = pnbp_values[0]
+            for value in pnbp_values[1:]:
+                smoothed = alpha * value + (1 - alpha) * smoothed
+            return max(0, smoothed)
+        else:
+            return np.mean(pnbp_values)
+
+
+class PNBPDataManager:
+    """Manajer data untuk PNBP predictions"""
+
+    @staticmethod
+    def load_historical_data(filepath="pnbp_history.json"):
+        """Load historical PNBP data from JSON file"""
+        try:
+            if os.path.exists(filepath):
+                with open(filepath, 'r') as f:
+                    return json.load(f)
+            else:
+                return []
+        except Exception as e:
+            print(f"Error loading historical data: {e}")
+            return []
+
+    @staticmethod
+    def save_predictions_to_json(predictions_data, filename="pnbp_predictions.json"):
+        """Menyimpan hasil prediksi PNBP ke file JSON"""
+        timestamp = datetime.now().strftime('%Y-%m-%d %H:%M:%S')
+
+        data_to_save = {
+            'timestamp': timestamp,
+            'predictions': predictions_data
+        }
+
+        try:
+            with open(filename, 'w') as f:
+                json.dump(data_to_save, f, indent=2, default=str)
+            return True
+        except Exception as e:
+            print(f"Error saving predictions: {e}")
+            return False
+
+    @staticmethod
+    def get_django_historical_data():
+        """Get historical PNBP data from Django models"""
+        try:
+            # Get data from CatchDetail model
+            catch_details = CatchDetail.objects.all().order_by('fish_catch__catch_date')
+            historical_data = []
+
+            for detail in catch_details:
+                if detail.pnbp and detail.pnbp > 0:
+                    historical_data.append({
+                        'date': detail.fish_catch.catch_date.isoformat(),
+                        'total_volume': float(detail.quantity),
+                        'total_pnbp': float(detail.pnbp),
+                        'gt_kapal': float(detail.fish_catch.ship.gross_tonnage) if detail.fish_catch.ship.gross_tonnage else 60
+                    })
+
+            return historical_data
+        except Exception as e:
+            print(f"Error getting Django historical data: {e}")
+            return []
+
+
+class PNBPredictor:
+    """Main class untuk menjalankan prediksi PNBP"""
+
+    def __init__(self):
+        self.regression_predictor = PNBPRegressionPredictor()
+        self.neural_predictor = PNBPNeuralNetworkPredictor()
+        self.optimization_predictor = PNBPOptimizationPredictor()
+        self.time_series_predictor = PNBPTimeSeriesPredictor()
+        self.data_manager = PNBPDataManager()
+
+    def run_prediction_pipeline(self, historical_data=None, ships_data=None):
+        """
+        Menjalankan pipeline prediksi PNBP lengkap dengan 4 metode.
+        """
+        if historical_data is None:
+            historical_data = self.data_manager.load_historical_data()
+
+        predictions = {}
+
+        # 1. Regression Analysis
+        predictions['regression'] = self.regression_predictor.predict(historical_data)
+
+        # 2. Neural Network
+        predictions['neural_network'] = self.neural_predictor.predict(historical_data)
+
+        # 3. NSGA-III Optimization
+        predictions['nsga3_optimization'] = self.optimization_predictor.predict(historical_data, ships_data)
+
+        # 4. Time Series Forecasting
+        predictions['time_series'] = self.time_series_predictor.predict(historical_data)
+
+        # 5. Normalize results
+        normalized_predictions = self._normalize_predictions(predictions)
+
+        # 6. Calculate weighted average
+        final_prediction = self._calculate_weighted_average(normalized_predictions)
+
+        return {
+            'individual_predictions': predictions,
+            'normalized_predictions': normalized_predictions,
+            'final_prediction': final_prediction,
+            'timestamp': datetime.now().isoformat(),
+            'methods_used': ['regression', 'neural_network', 'nsga3_optimization', 'time_series']
+        }
+
+    def _normalize_predictions(self, predictions):
+        """Normalisasi hasil prediksi menggunakan MinMaxScaler"""
+        if not predictions:
+            return {}
+
+        values = list(predictions.values())
+        scaler = MinMaxScaler()
+        normalized_values = scaler.fit_transform(np.array(values).reshape(-1, 1)).flatten()
+
+        return dict(zip(predictions.keys(), normalized_values))
+
+    def _calculate_weighted_average(self, normalized_predictions):
+        """Menghitung rata-rata tertimbang dari prediksi ternormalisasi"""
+        weights = {
+            'regression': 0.25,
+            'neural_network': 0.35,
+            'nsga3_optimization': 0.25,
+            'time_series': 0.15
+        }
+
+        weighted_sum = 0
+        total_weight = 0
+
+        for method, prediction in normalized_predictions.items():
+            if method in weights:
+                weighted_sum += prediction * weights[method]
+                total_weight += weights[method]
+
+        if total_weight == 0:
+            return 0
+
+        return weighted_sum / total_weight
+
+    def save_results(self, predictions_data, filename="pnbp_predictions.json"):
+        """Menyimpan hasil prediksi ke file JSON"""
+        return self.data_manager.save_predictions_to_json(predictions_data, filename)
+
+    def get_historical_data_from_django(self):
+        """Mendapatkan data historis dari model Django"""
+        return self.data_manager.get_django_historical_data()
+
+
+# -----------------------------------------------------------------------------------
 # PNBP Prediction API Views
 # -----------------------------------------------------------------------------------
 
@@ -570,12 +922,6 @@ def predict_pnbp_future(request):
     """
     API endpoint untuk prediksi PNBP masa depan menggunakan 4 metode analisis.
     """
-    if PNBPredictor is None:
-        return Response(
-            {'error': 'PNBP prediction module not available'},
-            status=status.HTTP_503_SERVICE_UNAVAILABLE
-        )
-
     try:
         # Initialize predictor
         predictor = PNBPredictor()
@@ -625,12 +971,6 @@ def get_pnbp_history(request):
     """
     API endpoint untuk mendapatkan data historis PNBP.
     """
-    if PNBPredictor is None:
-        return Response(
-            {'error': 'PNBP prediction module not available'},
-            status=status.HTTP_503_SERVICE_UNAVAILABLE
-        )
-
     try:
         predictor = PNBPredictor()
